fastapi-app/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Debsourya Datta <91617309+debsouryadatta@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.112.1"
uvicorn = "^0.30.6"
sqlmodel = "^0.0.21"
python-dotenv = "^1.0.1"
pydantic = "^2.8.2"
typing-extensions = "^4.12.2"
langchain = "^0.2.14"
langchain-openai = "^0.1.22"
openai = "^1.42.0"
youtube-transcript-api = "^0.6.2"
requests = "^2.32.3"
langchain-groq = "^0.1.9"
python-multipart = "^0.0.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
